rpk: separate version from the version command
We want to consume the version that is injected
at build time in other parts of the codebase
and we don't want to tie the package to the
command, so we introduce a separate package
for that.
