+++ src/go/rpk/pkg/cli/root.go
@@ -35,6 +35,7 @@ import (
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/topic"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/transform"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/version"
+	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cli/version/versioncmd"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/cobraext"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/config"
 	"github.com/redpanda-data/redpanda/src/go/rpk/pkg/plugin"
@@ -121,7 +122,7 @@ func Execute() {
 		security.NewCommand(fs, p),
 		topic.NewCommand(fs, p),
 		transform.NewCommand(fs, p, osExec),
-		version.NewCommand(fs, p),
+		versioncmd.NewCommand(fs, p),
 
 		newStatusCommand(), // deprecated
 	)
